+++ src/controllers/Parameter.ts
@@ -4,7 +4,7 @@
 import { CONTROLLER_METHOD_PARAMETERS } from "../constants";
 import { ParameterOptions } from "../types";
 import type { IControllerMethodParameter, Nilable } from "../types/internal";
-import { isNil } from "../utils";
+import { getFunctionParamNames, isNil } from "../utils";
 import { getListFromObject } from "./utils";
 
 /**
@@ -41,8 +41,8 @@ export function Parameter(options?: Nilable<ParameterOptions>): ParameterDecorat
         throw new TypeError("options must be of type object");
     }
 
-    if (!isNil(options.name)) {
-        if (typeof options.name !== "string") {
+    if (!isNil((options as any).name)) {
+        if (typeof (options as any).name !== "string") {
             throw new TypeError("options.name must be of type string");
         }
     }
@@ -50,9 +50,9 @@ export function Parameter(options?: Nilable<ParameterOptions>): ParameterDecorat
     return function (target, propertyKey, parameterIndex) {
         const method: Function = (target as any)[propertyKey];
 
-        let parameterName = options!.name;
+        let parameterName = (options as any).name;
         if (!parameterName?.trim().length) {
-            parameterName = getParamNames(method)[parameterIndex];
+            parameterName = getFunctionParamNames(method)[parameterIndex];
         }
 
         if (!parameterName?.trim().length) {
@@ -68,22 +68,4 @@ export function Parameter(options?: Nilable<ParameterOptions>): ParameterDecorat
             }
         );
     };
-}
-
-// s. https://stackoverflow.com/questions/1007981/how-to-get-function-parameter-names-values-dynamically
-function getParamNames(func: Function) {
-    let result: Nilable<any[]>;
-
-    try {
-        const STRIP_COMMENTS = /((\/\/.*$)|(\/\*[\s\S]*?\*\/))/mg;
-        const ARGUMENT_NAMES = /([^\s,]+)/g;
-
-        const fnStr = func.toString().replace(STRIP_COMMENTS, "");
-        result = fnStr.slice(fnStr.indexOf("(") + 1, fnStr.indexOf(")")).match(ARGUMENT_NAMES);
-    }
-    catch {
-        result = undefined;
-    }
-
-    return result ?? [];
 }
